Refactor neural-network lib to individual files
