+++ libs/neural-network/src/layer_topology.rs
@@ -1,0 +1,3 @@
+pub struct LayerTopology {
+    pub neurons: usize,
+}
